requirements/local.txt
asgiref==3.8.1
Django==5.1.1
django-environ==0.11.2
psycopg2==2.9.9
sqlparse==0.5.1
tzdata==2024.2
pillow==10.4.0
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3